+++ core/ajax/gsl.ajax.php
@@ -50,6 +50,14 @@ try {
 		gsl::google_logout();
 		ajax::success();
 	}
+  
+  if (init('action') == 'cookie') {
+		if (!isConnect('admin')) {
+			throw new Exception(__('401 - Accès non autorisé', __FILE__));
+		}
+		gsl::google_saveCookie(init('cookie'));
+		ajax::success();
+	}
 
 	throw new Exception(__('Aucune méthode correspondante à : ', __FILE__) . init('action'));
 	/*     * *********Catch exeption*************** */